integrated payment conf page, modified scanner page
