+++ superpark-frontend/components/DrawerScreens.js
@@ -22,6 +22,7 @@ const DrawerScreens = () => {
                   />
                 )
               }}
+              initialParams={{headerShown: true}}
             />
           ))
         }
