+++ superpark-frontend/package.json
@@ -16,6 +16,7 @@
     "@react-navigation/stack": "^7.1.1",
     "depcheck": "^1.4.7",
     "expo": "~52.0.33",
+    "expo-camera": "^16.0.17",
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
